+++ capytaine/bodies/bodies.py
@@ -94,8 +94,8 @@ class FloatingBody(Abstract3DObject):
                 triangles_as_quads[:, 3] = cells['triangle'][:, 2]  # Repeat one node to make a quad
                 all_faces.append(triangles_as_quads)
             return np.concatenate(all_faces)
-        
-        cpt_mesh = Mesh(vertices=mesh.points, 
+
+        cpt_mesh = Mesh(vertices=mesh.points,
                         faces=all_faces_as_quads(mesh.cells_dict),
                         name=name+"_mesh")
 
@@ -230,6 +230,431 @@ class FloatingBody(Abstract3DObject):
         return xr.DataArray(data=np.asarray(matrix), dims=['influenced_dof', 'radiating_dof'],
                             coords={'influenced_dof': list(self.dofs), 'radiating_dof': list(self.dofs)},
                             )
+
+    ###################
+    # Hydrostatics #
+    ###################
+
+    def surface_integral(self, data, **kwargs):
+        """Returns integral of given data along wet surface area."""
+        return np.sum(data * self.mesh.faces_areas, **kwargs)
+
+    def waterplane_integral(self, data, **kwargs):
+        """Returns integral of given data along water plane area."""
+        return self.surface_integral(self.mesh.faces_normals[:,2] * data, **kwargs)
+
+    @property
+    def wet_surface_area(self):
+        """Returns wet surface area."""
+        return self.surface_integral(1)
+
+    @property
+    def volumes(self):
+        """Returns volumes using x, y, z components of the FloatingBody."""
+        norm_coord = self.mesh.faces_normals * self.mesh.faces_centers
+        return self.surface_integral(norm_coord.T, axis=1)
+
+    @property
+    def volume(self):
+        """Returns volume of the FloatingBody."""
+        return np.mean(self.volumes)
+
+    def disp_mass(self, *, rho=1000):
+        return rho * self.volume
+
+    @property
+    def center_of_buoyancy(self):
+        """Returns center of buoyancy of the FloatingBody."""
+        volume = self.volume
+        coords_sq_norm = self.mesh.faces_normals * self.mesh.faces_centers**2
+        return self.surface_integral(coords_sq_norm.T, axis=1) / (2*volume)
+
+    @property
+    def waterplane_area(self):
+        """Returns water plane area of the FloatingBody."""
+        waterplane_area = -self.waterplane_integral(1)
+        return waterplane_area
+
+    @property
+    def waterplane_center(self):
+        """Returns water plane center of the FloatingBody.
+
+        Note: Returns None if the FloatingBody is full submerged.
+        """
+        waterplane_area = self.waterplane_area
+        if abs(waterplane_area) < 1e-10:
+            return None
+        else:
+            waterplane_center = -self.waterplane_integral(
+                self.mesh.faces_centers.T, axis=1) / waterplane_area
+            return waterplane_center[:-1]
+
+    @property
+    def transversal_metacentric_radius(self):
+        """Returns transversal metacentric radius of the body."""
+        inertia_moment = -self.waterplane_integral(self.mesh.faces_centers[:,1]**2)
+        return inertia_moment / self.volume
+
+    @property
+    def longitudinal_metacentric_radius(self):
+        """Returns longitudinal metacentric radius of the body."""
+        inertia_moment = -self.waterplane_integral(self.mesh.faces_centers[:,0]**2)
+        return inertia_moment / self.volume
+
+    @property
+    def transversal_metacentric_height(self):
+        """Returns transversal metacentric height of the body."""
+        gb = self.center_of_mass - self.center_of_buoyancy
+        return self.transversal_metacentric_radius - gb[2]
+
+    @property
+    def longitudinal_metacentric_height(self):
+        """Returns longitudinal metacentric height of the body."""
+        gb = self.center_of_mass - self.center_of_buoyancy
+        return self.longitudinal_metacentric_radius - gb[2]
+
+    def dof_normals(self, dof):
+        """Returns dot product of the surface face normals and DOF"""
+        return np.sum(self.mesh.faces_normals * dof, axis=1)
+
+    def each_hydrostatic_stiffness(self, influenced_dof_name, radiating_dof_name, *,
+                                         influenced_dof_div=0.0, rho=1000.0, g=9.81):
+        r"""
+        Return the hydrostatic stiffness for a pair of DOFs.
+
+        :math:`C_{ij} = \rho g\iint_S (\hat{n} \cdot V_j) (w_i + z D_i) dS`
+
+        where :math:`\hat{n}` is surface normal,
+
+        :math:`V_i = u_i \hat{n}_x + v_i \hat{n}_y + w_i \hat{n}_z` is DOF vector and
+
+        :math:`D_i = \nabla \cdot V_i` is the divergence of the DOF.
+
+        Parameters
+        ----------
+        influenced_dof_name : str
+            Name of influenced DOF vector of the FloatingBody
+        radiating_dof_name: str
+            Name of radiating DOF vector of the FloatingBody
+        influenced_dof_div: np.ndarray (Face_count), optional
+            Influenced DOF divergence of the FloatingBody, by default 0.0.
+        rho: float, optional
+            water density, by default 1000.0
+        g: float, optional
+            Gravity acceleration, by default 9.81
+
+        Returns
+        -------
+        hs_ij: xarray.variable
+            hydrostatic_stiffness of ith DOF and jth DOF.
+
+        Note
+        ----
+            This function computes the hydrostatic stiffness assuming :math:`D_{i} = 0`.
+            If :math:`D_i \neq 0`, input the divergence interpolated to face centers.
+
+            General integral equations are used for the rigid body modes and
+            Neumann (1994) method is used for flexible modes.
+
+        References
+        ----------
+            Newman, John Nicholas. "Wave effects on deformable bodies."Applied ocean
+            research" 16.1 (1994): 47-59.
+            http://resolver.tudelft.nl/uuid:0adff84c-43c7-43aa-8cd8-d4c44240bed8
+
+        """
+        # Newmann (1994) formula is not 'complete' as recovering the rigid body
+        # terms is not possible. https://doi.org/10.1115/1.3058702.
+
+        # Alternative is to use the general equation of hydrostatic and
+        # restoring coefficient for rigid mdoes and use Neuman equation for elastic
+        # modes.
+        cog = self.center_of_mass
+
+        rigid_dof_names = ("Surge", "Sway", "Heave", "Roll", "Pitch", "Yaw")
+
+        dof_pair = (influenced_dof_name, radiating_dof_name)
+
+        if set(dof_pair).issubset(set(rigid_dof_names)):
+            if dof_pair == ("Heave", "Heave"):
+                norm_hs_stiff = self.waterplane_area
+            elif dof_pair in [("Heave", "Roll"), ("Roll", "Heave")]:
+                norm_hs_stiff = -self.waterplane_integral(self.mesh.faces_centers[:,1])
+            elif dof_pair in [("Heave", "Pitch"), ("Pitch", "Heave")]:
+                norm_hs_stiff = self.waterplane_integral(self.mesh.faces_centers[:,0])
+            elif dof_pair == ("Roll", "Roll"):
+                norm_hs_stiff = self.volume * self.transversal_metacentric_height
+            elif dof_pair in [("Roll", "Pitch"), ("Pitch", "Roll")]:
+                norm_hs_stiff = self.waterplane_integral(self.mesh.faces_centers[:,0]
+                                                          * self.mesh.faces_centers[:,1])
+            elif dof_pair == ("Roll", "Yaw"):
+                norm_hs_stiff = self.volume * (-self.center_of_buoyancy[0] + cog[0])
+            elif dof_pair == ("Pitch", "Pitch"):
+                norm_hs_stiff = self.volume * self.longitudinal_metacentric_height
+            elif dof_pair == ("Pitch", "Yaw"):
+                norm_hs_stiff = self.volume * (-self.center_of_buoyancy[1] + cog[1])
+            else:
+                norm_hs_stiff = 0.0
+        else:
+            # Neuman (1994) formula for flexible DOFs
+            influenced_dof = np.array(self.dofs[influenced_dof_name])
+            radiating_dof = np.array(self.dofs[radiating_dof_name])
+            influenced_dof_div_array = np.array(influenced_dof_div)
+
+            radiating_dof_normal = self.dof_normals(radiating_dof)
+            z_influenced_dof_div = influenced_dof[:,2] + self.mesh.faces_centers[:,2] * influenced_dof_div_array
+            norm_hs_stiff = self.surface_integral( -radiating_dof_normal * z_influenced_dof_div)
+
+        hs_stiff = rho * g * norm_hs_stiff
+
+        return xr.DataArray([[hs_stiff]],
+                            dims=['influenced_dof', 'radiating_dof'],
+                            coords={'influenced_dof': [influenced_dof_name],
+                            'radiating_dof': [radiating_dof_name]},
+                            name="hydrostatic_stiffness"
+                            )
+
+    def compute_hydrostatic_stiffness(self, *, divergence=None, rho=1000.0, g=9.81):
+        r"""
+        Compute hydrostatic stiffness matrix for all DOFs of the body.
+
+        :math:`C_{ij} = \rho g\iint_S (\hat{n} \cdot V_j) (w_i + z D_i) dS`
+
+        where :math:`\hat{n}` is surface normal,
+
+        :math:`V_i = u_i \hat{n}_x + v_i \hat{n}_y + w_i \hat{n}_z` is DOF vector and
+
+        :math:`D_i = \nabla \cdot V_i` is the divergence of the DOF.
+
+        Parameters
+        ----------
+        divergence : dict mapping a dof name to an array of shape (nb_faces) or
+                        xarray.DataArray of shape (nb_dofs × nb_faces), optional
+            Divergence of the DOFs, by default None
+        rho : float, optional
+            Water density, by default 1000.0
+        g: float, optional
+            Gravity acceleration, by default 9.81
+
+        Returns
+        -------
+        xr.DataArray
+            Matrix of hydrostatic stiffness
+
+        Note
+        ----
+            This function computes the hydrostatic stiffness assuming :math:`D_{i} = 0`.
+            If :math:`D_i \neq 0`, input the divergence interpolated to face centers.
+
+            General integral equations are used for the rigid body modes and
+            Neumann (1994) method is used for flexible modes.
+
+        References
+        ----------
+            Newman, John Nicholas. "Wave effects on deformable bodies."Applied ocean
+            research" 16.1 (1994): 47-59.
+            http://resolver.tudelft.nl/uuid:0adff84c-43c7-43aa-8cd8-d4c44240bed8
+
+        """
+        if len(self.dofs) == 0:
+            raise AttributeError("Cannot compute hydrostatics stiffness on {} since no dof has been defined.".format(self.name))
+
+        def divergence_dof(influenced_dof):
+            if divergence is None:
+                return 0.0
+            elif isinstance(divergence, dict) and influenced_dof in divergence.keys():
+                return divergence[influenced_dof]
+            elif isinstance(divergence, xr.DataArray) and influenced_dof in divergence.coords["influenced_dof"]:
+                return divergence.sel(influenced_dof=influenced_dof).values
+            else:
+                LOG.warning("Computing hydrostatic stiffness without the divergence of {}".format(influenced_dof))
+                return 0.0
+
+        hs_set =  xr.merge([
+            self.each_hydrostatic_stiffness(
+                influenced_dof_name, radiating_dof_name,
+                influenced_dof_div = divergence_dof(influenced_dof_name),
+                rho=rho, g=g
+                )
+            for radiating_dof_name in self.dofs
+            for influenced_dof_name in self.dofs
+            ])
+
+        return hs_set.hydrostatic_stiffness
+
+    def compute_rigid_body_inertia(self, *, rho=1000, output_type="body_dofs"):
+        """
+        Inertia Mass matrix of the body for 6 rigid DOFs.
+
+        Parameters
+        ----------
+        rho : float, optional
+            Density of water, by default 1000.0
+        output_type : {"body_dofs", "rigid_dofs", "all_dofs"}
+            Type of DOFs for mass mat output, by default "body_dofs".
+
+        Returns
+        -------
+        xarray.DataArray
+            Inertia matrix
+
+        Raises
+        ------
+        ValueError
+            If output_type is not in {"body_dofs", "rigid_dofs", "all_dofs"}.
+        """
+
+        cog = self.center_of_mass
+
+        fcs = (self.mesh.faces_centers).T
+        combinations = np.array([fcs[0]**2, fcs[1]**2, fcs[2]**2, fcs[0]*fcs[1],
+                                 fcs[1]*fcs[2], fcs[2]*fcs[0]])
+        integrals = np.array([
+            [np.sum(normal_i * fcs[axis] * combination * self.mesh.faces_areas)
+            for combination in combinations]
+            for axis, normal_i in enumerate(self.mesh.faces_normals.T)])
+
+        inertias = rho * np.array([
+            (integrals[0,1]   + integrals[0,2]   + integrals[1,1]/3 \
+             + integrals[1,2]   + integrals[2,1] + integrals[2,2]/3)/3,
+            (integrals[0,0]/3 + integrals[0,2]   + integrals[1,0]   \
+             + integrals[1,2]   + integrals[2,0] + integrals[2,2]/3)/3,
+            (integrals[0,0]/3 + integrals[0,1]   + integrals[1,0]   \
+             + integrals[1,1]/3 + integrals[2,0] + integrals[2,1]  )/3,
+            integrals[2,3],
+            integrals[0,4],
+            integrals[1,5]
+        ])
+
+        mass = self.disp_mass(rho=rho)
+        mass_mat = np.array([
+            [ mass       ,  0          ,  0           ,
+              0          ,  mass*cog[2], -mass*cog[1]],
+            [ 0          ,  mass       ,  0           ,
+             -mass*cog[2],  0          ,  mass*cog[0]],
+            [ 0          ,  0          ,  mass        ,
+              mass*cog[1], -mass*cog[0],  0          ],
+            [ 0          , -mass*cog[2],  mass*cog[1] ,
+              inertias[0], -inertias[3], -inertias[5]],
+            [ mass*cog[2],  0          , -mass*cog[0] ,
+             -inertias[3],  inertias[1], -inertias[4]],
+            [-mass*cog[1],  mass*cog[0],  0           ,
+             -inertias[5], -inertias[4], inertias[2]] ,
+        ])
+
+        # Rigid DOFs
+        rigid_dof_names = ["Surge", "Sway", "Heave", "Roll", "Pitch", "Yaw"]
+        rigid_mass_xr = xr.DataArray(data=np.asarray(mass_mat),
+                            dims=['influenced_dof', 'radiating_dof'],
+                            coords={'influenced_dof': rigid_dof_names,
+                                    'radiating_dof': rigid_dof_names},
+                            name="mass")
+
+        # Body DOFs (Default as np.nan)
+        body_dof_names = list(self.dofs)
+        body_dof_count = len(body_dof_names)
+        body_mass_xr = xr.DataArray(np.nan * np.zeros([body_dof_count, body_dof_count]),
+                                    dims=['influenced_dof', 'radiating_dof'],
+                                    coords={'influenced_dof': body_dof_names,
+                                            'radiating_dof': body_dof_names},
+                                    name="mass")
+
+        total_mass_xr = xr.merge([rigid_mass_xr, body_mass_xr], compat="override").mass
+
+        non_rigid_dofs = set(body_dof_names) - set(rigid_dof_names)
+
+        if output_type == "body_dofs":
+            if len(non_rigid_dofs) > 0:
+                LOG.warning(f"Non-rigid dofs: {non_rigid_dofs} are detected and \
+respective inertia coefficients are assigned as NaN.")
+
+            mass_xr = total_mass_xr.sel(influenced_dof=body_dof_names,
+                                            radiating_dof=body_dof_names)
+        elif output_type == "rigid_dofs":
+            mass_xr = total_mass_xr.sel(influenced_dof=rigid_dof_names,
+                                        radiating_dof=rigid_dof_names)
+        elif output_type == "all_dofs":
+            if len(non_rigid_dofs) > 0:
+                LOG.warning("Non-rigid dofs: {non_rigid_dofs} are detected and \
+respective inertia coefficients are assigned as NaN.")
+
+            mass_xr = total_mass_xr
+        else:
+            raise ValueError(f"output_type should be either 'body_dofs', \
+'all_dofs' or 'rigid_dofs'. Given output_type = '{output_type}'.")
+
+        return mass_xr
+
+
+    def compute_hydrostatics(self, *, rho=1000.0, g=9.81, divergence=None):
+        """Compute hydrostatics of the FloatingBody.
+
+        Parameters
+        ----------
+        rho : float, optional
+            Density of Water. The default is 1000.
+        g: float, optional
+            Gravity acceleration. The default is 9.81.
+        divergence : np.ndarray, optional
+            Divergence of the DOFs.
+
+        Returns
+        -------
+        hydrostatics : dict
+            All hydrostatics values of the FloatingBody.
+        """
+        if not hasattr(self, "center_of_mass"):
+            LOG.warning("The floating body {} has no defined center of mass. The center of mass is set to (0,0,0).".format(self.name))
+            self.center_of_mass = np.array([0.0, 0.0, 0.0])
+
+        self.keep_immersed_part()
+
+        full_mesh_vertices = self.full_body.mesh.vertices
+        coord_max = full_mesh_vertices.max(axis=0)
+        coord_min = full_mesh_vertices.min(axis=0)
+        full_length, full_breadth, depth = full_mesh_vertices.max(axis=0) - full_mesh_vertices.min(axis=0)
+
+        vertices = self.mesh.vertices
+        sub_length, sub_breadth, _ = vertices.max(axis=0) - vertices.min(axis=0)
+
+        if abs(self.waterplane_area) > 1e-10:
+            water_plane_idx = np.isclose(vertices[:,2], 0.0)
+            water_plane = vertices[water_plane_idx][:,:-1]
+            wl_length, wl_breadth = water_plane.max(axis=0) - water_plane.min(axis=0)
+        else:
+            wl_length, wl_breadth = 0.0, 0.0
+
+        hydrostatics = {}
+        hydrostatics["g"] = g
+        hydrostatics["rho"] = rho
+        hydrostatics["center_of_mass"] = self.center_of_mass
+
+        hydrostatics["wet_surface_area"] = self.wet_surface_area
+        hydrostatics["disp_volumes"] = self.volumes
+        hydrostatics["disp_volume"] = self.volume
+        hydrostatics["disp_mass"] = self.disp_mass(rho=rho)
+        hydrostatics["center_of_buoyancy"] = self.center_of_buoyancy
+        hydrostatics["waterplane_center"] = np.append(self.waterplane_center, 0.0)
+        hydrostatics["waterplane_area"] = self.waterplane_area
+        hydrostatics["transversal_metacentric_radius"] = self.transversal_metacentric_radius
+        hydrostatics["longitudinal_metacentric_radius"] = self.longitudinal_metacentric_radius
+        hydrostatics["transversal_metacentric_height"] = self.transversal_metacentric_height
+        hydrostatics["longitudinal_metacentric_height"] = self.longitudinal_metacentric_height
+        hydrostatics["hydrostatic_stiffness"] = self.compute_hydrostatic_stiffness(
+            divergence=divergence, rho=rho, g=g)
+
+        hydrostatics["length_overall"] = full_length
+        hydrostatics["breadth_overall"] = full_breadth
+        hydrostatics["depth"] = depth
+        hydrostatics["draught"] = np.abs(coord_min[2])
+        hydrostatics["length_at_waterline"] = wl_length
+        hydrostatics["breadth_at_waterline"] = wl_breadth
+        hydrostatics["length_overall_submerged"] = sub_length
+        hydrostatics["breadth_overall_submerged"] = sub_breadth
+        hydrostatics["inertia_matrix"] = self.compute_rigid_body_inertia(rho=rho)
+
+        return hydrostatics
+
 
     ###################
     # Transformations #