+++ docs/changelog.rst
@@ -27,6 +27,8 @@ Changelog
 * The functions :code:`problems_from_dataset` and :code:`fill_dataset` accept a body alone as input.
   That is, one can use :code:`fill_dataset(test_matrix, body)` and not only :code:`fill_dataset(test_matrix, [body])` (pull:`144`).
 
+* Add method to compute some of the hydrostatic parameters such as volume, buoyancy center, wet surface area, hydrostatic stiffness, inertia matrix etc.
+  :code:`compute_hydrostatics` method is created to return all hydrostatic parameters similar to :code:`meshmagick.hydrostatics.compute_hydrostatics` (:pull:`106`).
 
 -------------------------------
 New in version 1.3 (2021-10-07)
@@ -47,6 +49,7 @@ Major changes
   :code:`reflection_symmetry` optional keyword arguments.
   The :code:`clever` keyword argument is deprecated for :code:`HorizontalCylinder`
   and should be replaced by the new more explicit keyword arguments above.
+
 
 New features
 ~~~~~~~~~~~~